rename qbittorrentdto.go;and so on
