+++ plugins/inputs/qbittorrent/dto.go
